+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - Add Winter Cyclamen to Howling Peaks
 - Add Protea to Fragment Jungle
 - Hitting Fruit Blocks with a Projectile will now cause the block to break with the fruit dropping if it is fully grown
-- Fix Incompatibily with Organics Mod. Closes #164
+- Fix Incompatibily with Organics Mod
+- Clicking on Fruit Leaves with Fruit Item in creative will now place the fruit block below the leaves (Assuming there is air below)
 
 
 # 1.4.2